requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Structure parsing and analysis
biopython>=1.81
gemmi>=0.6.0

# Chemistry
rdkit>=2023.3.1

# Web visualization (3Dmol.js is CDN-based, no install needed)
# PyMOL is optional, install separately

# Utilities
requests>=2.31.0
tqdm>=4.65.0

# Optional: ESM embeddings
# fair-esm>=2.0.0
